Remove clone requirements.
I have a conundrum. The canonical implementation
of this will not require Clone trait on client T,
but I am unable to implement the in memory version
without it. The in memory version greatly simplifies
testing. Therefore I force the trait Clone on the client
in the name of testing. This commit compiles without
Clone trait, but tests will fail to compile as LogEntry
does not implement Clone.